# doc/api.yml
openapi: 3.0.3
info:
  title: Cinema API
  description: Booking API for a cinema - rooms, movies, projections and reservations.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /projections:
    get:
      tags: [ Projections ]
      summary: List all projections, open to everyone
      security: []
      responses:
        '200':
          description: Projections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Projection'
    put:
      tags: [ Projections ]
      summary: Create a projection (MANAGER only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectionRequest'
      responses:
        '201':
          description: Projection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projection'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: The room is already booked for that time slot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projections/{id}:
    get:
      tags: [ Projections ]
      summary: Get a projection by id
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projection'
        '404':
          $ref: '#/components/responses/NotFound'
  /projections/{id}/available-seats:
    get:
      tags: [ Projections ]
      summary: List the seats not yet booked for a projection, open to everyone
      security: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Available seats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Seat'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: The authenticated role is not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
...
    Seat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
...
    ProjectionRequest:
      type: object
      required: [ movieId, roomId, datetime, seatPrice ]
      properties:
        movieId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        datetime:
          type: string
          format: date-time
        seatPrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
    Projection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        datetime:
          type: string
          format: date-time
        seatPrice:
          type: number
        movieId:
          type: string
          format: uuid
        movieTitle:
          type: string
        roomId:
          type: string
          format: uuid
        roomNumber:
          type: string
...
security:
  - bearerAuth: []